// src/app/enums/enums.ts
export enum EtapeTraitement {
    SOUMISSION = 'SOUMISSION',
    RECEPTION = 'RECEPTION',
    VALIDATION = 'VALIDATION',
    IMPUTATION = 'IMPUTATION',
    TRAITEMENT='TRAITEMENT',
    CLOTURE = 'CLOTURE',
    PUBLISHED = 'PUBLISHED',
    VALIDATION_PILOTE = 'VALIDATION_PILOTE',
    VALIDATION_RS = 'VALIDATION_RS',
    SUIVI_RQ = 'SUIVI_RQ',
    VALIDATION_PLAN = 'VALIDATION_PLAN',


}
export enum StatusEnum {
    PENDIND = 'PENDIND',

    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
    IN_PROGRESS = 'IN_PROGRESS',

}
export enum TypeStructure {
    DIRECTION = 'DIRECTION',
    SERVICE = 'SERVICE'
}

export enum NonConformStatus {
    PENDIND = 'PENDIND',
    NON_TRAITER = 'NON_TRAITER',
    ARCHIVED = 'ARCHIVED',
    REJECTED = 'REJECTED',
    PUBLISHED = 'PUBLISHED',
    TRAITER = 'TRAITER',
    DRAFT = 'DRAFT',
    IN_PROGRESS = 'IN_PROGRESS',

}